Properly deal with version oddities.
RubyGems has a somewhat odd though understandable way of dealing with version
numbers. For example, it considers the version "3.4" the same as "3.4.0". While
this makes sense it gets a bit annoying when you want to download the `.gem`
file. In case of Slop version 3.4 gem-mirror would try to download
"slop-3.4.gem" while the file was actually supposed to be "slop-3.4.0.gem".

This commit addresses this issue by filtering out non existing versions and
using the latest version in case none were left.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>